Add proper multibyte UTF-8 support
libhyphen itself supports UTF-8, but the example code it comes with
uses single-byte tolower() presuming ASCII with a comment about how it
is not suitable for real-world use.

Implemented using towlower() which has the advantage of being a
standard library, but has the disadvantage that I had to convert to
widecharacters and back.